+++ vecorel_cli/cli/options.py
@@ -31,7 +31,16 @@ GEOPARQUET_COMPRESSION = click.option(
     type=click.Choice(COMPRESSION_METHODS),
     help="GeoParquet only: Compression method",
     show_default=True,
-    default="brotli",
+    default="zstd",
+)
+
+GEOPARQUET_COMPRESSION_LEVEL = click.option(
+    "--compression_level",
+    "-pcl",
+    type=click.IntRange(min=1, max=22),
+    help="GeoParquet only: Compression level",
+    show_default=True,
+    default=None,
 )
 
 GEOPARQUET_VERSION = click.option(